dev.requirements.txt
-r requirements.txt

pytest
pytest-ordering
pytest-asyncio==0.18.3
sqlalchemy2-stubs
testcontainers[postgresql]
